refactor: move linechart feat into linechart_helper.py
